// ssyweb/SSYNEW/src/app/claim/models/attachment.model.ts
export class AttachmentModel {
    attachmentId: number;
    attachmentType: number;
    claimType: number;
    transactionId: number;
    fileName: string;
    filePath: string;
    fileContent: any;
    statusId: number;
    createdBy: number;
    attachmenturl: any;
    stringContent: string;
    attachmentTypeString: string;
}